bug: fix management of IPv4 route configuration file
This addresses a regression in the handling of
/etc/sysconfig/network-scripts/route-* introduced in version 1.6.1
with commit 4b3d2f587 ("Move route file creation to a dhclient enter
hook")

Prior to that commit, ec2ifup would trigger the removal of the route
config file, but otherwise that file would be preserved.  The
regression introduced by that commit was that the route file would be
overwritten any time `ifup` brought the interface up, not only by
`ec2ifup`.

Fixes https://github.com/amazonlinux/amazon-ec2-net-utils/issues/82

+++ 50_ec2_rewrite_primary_enter_hook.sh
@@ -61,16 +61,18 @@ ec2_rewrite_primary_enter_hook() {
     [ -z "$gateway" ] && { logger --tag ec2net "DHCP lease did not set gateway" ; return 1; }
     [ -z "$route_dest" ] && { logger --tag ec2net "DHCP lease did not set route_dest" ; return 1; }
 
-    if subnet_supports_ipv4 "$new_ip_address"; then
-        cat << EOF > ${route_file}
+    if [ ! -f "${route_file}" ]; then
+        if subnet_supports_ipv4 "$new_ip_address"; then
+            cat << EOF > ${route_file}
 default via ${gateway} dev ${interface} table ${RTABLE}
 ${route_dest} dev ${interface} proto kernel scope link src ${new_ip_address} table ${RTABLE}
 EOF
-        if should_use_mainroutetable "$config_file"; then
-            logger --tag ec2net "[dhclient] adding default route to main table for ${interface} metric ${RTABLE}"
-            cat << EOF >> ${route_file}
+            if should_use_mainroutetable "$config_file"; then
+                logger --tag ec2net "[dhclient] adding default route to main table for ${interface} metric ${RTABLE}"
+                cat << EOF >> ${route_file}
 default via ${gateway} dev ${interface} metric ${RTABLE}
 EOF
+            fi
         fi
     fi
 }
